README.md
Agenda: 
The code of this website is for absolute beginners. The basic concepts of HTML and CSS are implemented in this website. You can see how to make your website responsive and effective at the same time at the very beginning stage of your coding journey.
Instructions for installing my code:
You can simply install my code through my git hub repository. You just have to go to the raw data and from there you can access the code and material.
Other code needs to be installed for it to work:
Right now I have made the initial interface this parallax website but there is much more to add in it to make it responsive as well as interactive. Code of CSS and JavaScript is be used for further enhancement.
How can you contribute: 
You can simply contribute by adding your ideas and code to my git hub account.